+++ src/super_gradients/modules/detection_modules.py
@@ -9,7 +9,7 @@ from omegaconf import DictConfig
 from super_gradients.training.models.classification_models.mobilenetv2 import InvertedResidual
 from super_gradients.training.utils.utils import HpmStruct
 from super_gradients.training.models import MobileNet, MobileNetV2
-from super_gradients.training.utils.module_utils import MultiOutputModule
+from super_gradients.modules.multi_output_modules import MultiOutputModule
 
 
 class BaseDetectionModule(nn.Module, ABC):
