+++ src/super_gradients/training/models/segmentation_models/context_modules.py
@@ -6,7 +6,7 @@ import torch
 import torch.nn.functional as F
 
 from super_gradients.modules import ConvBNReLU
-from super_gradients.training.utils.module_utils import UpsampleMode
+from super_gradients.modules.sampling import UpsampleMode
 from super_gradients.common.object_names import ContextModules
