+++ src/super_gradients/training/models/segmentation_models/ppliteseg.py
@@ -2,7 +2,7 @@ import torch
 import torch.nn as nn
 from typing import Union, List
 from super_gradients.modules import ConvBNReLU
-from super_gradients.training.utils.module_utils import make_upsample_module
+from super_gradients.modules.sampling import make_upsample_module
 from super_gradients.common.data_types.enum.upsample_mode import UpsampleMode
 from super_gradients.training.models.segmentation_models.stdc import AbstractSTDCBackbone, STDC1Backbone, STDC2Backbone
 from super_gradients.training.models.segmentation_models.common import SegmentationHead
