+++ src/super_gradients/training/models/segmentation_models/unet/unet.py
@@ -4,7 +4,9 @@ from typing import Optional, Union, List
 from super_gradients.training.utils import HpmStruct, get_param
 from super_gradients.training import models
 from super_gradients.training.models.segmentation_models.segmentation_module import SegmentationModule
-from super_gradients.training.utils.module_utils import make_upsample_module, UpsampleMode, fuse_repvgg_blocks_residual_branches
+from super_gradients.modules.sampling import UpsampleMode
+from super_gradients.modules.sampling import make_upsample_module
+from super_gradients.modules.repvgg_block import fuse_repvgg_blocks_residual_branches
 from super_gradients.training.models.segmentation_models.unet.unet_encoder import UNetBackboneBase, Encoder
 from super_gradients.training.models.segmentation_models.context_modules import AbstractContextModule
 from super_gradients.training.models.segmentation_models.unet.unet_decoder import Decoder
